chore: 🚧 Update openai.ts file
Update the getBasePrompt function in openai.ts file to remove the INSTRUCTIONS header and add the language of the commit. Update the getCommitMessageExtraContext function to add guidelines for the commit title and description. The commit title should be no longer than 50 characters and the commit description should be no longer than 72 characters. The commit message should be in the following format: <conventional commits type>(<optional scope of the change>): <gitmoji> <commit title>/n/n<commit description>. This commit includes changes to improve the prompt and formatting of the commit message.